Change directory path from java to Java
