--- src/main/java/dukeychatbot/dukeyexceptions/InvalidCommandException.java
@@ -1,19 +1,0 @@
-package dukeychatbot.dukeyexceptions;
-
-/**
- * Constructs InvalidCommandException error which inherits from DukeyException.
- *
- * <p>Returns error when user inputs a command is not provided.</p>
- */
-public class InvalidCommandException extends DukeyException {
-
-    public InvalidCommandException() {
-        super("____________________________________________________________\n" +
-              "WARNING: Invalid command. \n" +
-              "Valid commands are: 1) todo, 2) deadline, 3) event to add commands. " +
-              "Type them to see their functionality\n" +
-              "To mark tasks as done or undone, use commands: 'mark <task number>' and 'unmark <task number>'\n" +
-              "To delete tasks, use command: 'delete <task number>'\n" +
-              "____________________________________________________________\n");
-    }
-}
